Make table-config an admin-only REST endpoint
Move GET /api/table-config out of the public client router into the admin
router as GET /api/admin/table-config, so a table's storage format and
configuration (which can include external SQL-database connection details) is
only readable with admin basic-auth.

- Add beacon-api/src/axum/admin/tables.rs with the list_table_config handler
  (admin tag, basic-auth, 404 on missing table) and register it.
- Remove the handler and route from the client surface; tidy tag descriptions.
- Keep credential redaction in TableConfigView as defense-in-depth; reword its
  now-inaccurate "public endpoint" comments.
- Update the integration test to call the admin path with auth and assert an
  unauthenticated call is rejected with 401.
- Update the VitePress REST API docs: move table-config under admin, document
  the full admin surface with example requests, and flag deprecated aliases.
